requirements.txt
pytorch_pretrained_bert==0.6.2
tqdm==4.65.0
numpy==1.24.1
transformers==4.27.1
yaml==0.2.5
pandas==2.0.3
openai==0.28.0
tiktoken==0.4.0
